Add a bunch to name lexer in preparation for fixing
The old simple strategy I was using to lex names was not going to ever work for non-literal words that both started and ended with curly braces (like "{\e'}...{\e'}", some word that starts and ends with special tex-characters). So here's a new strategy which should adapt to that, but is super ugly - not sure how to fix it without getting really verbose and tokenizing it properly.