Fixed initials parsing in leading names with whitespace
Authors with initials were not being parsed right if they had whitespace in the name and came first in the list. This is because the first rule of the name-lexer had fragments and initials in the wrong order - if there was whitespace present, the fragment rule would match and gobble up the initials. Names that come later are not processed by the same rule and so were unaffected.